+++ Packages/com.l47coder.dev-workbench/Runtime~/Templates/Managers/Component/ComponentManager.cs
@@ -12,7 +12,10 @@ public interface IComponentManager
 
 internal sealed partial class ComponentManagerData
 {
+    [TableColumn(Editable = false)]
     public string Key;
+
+    [TableColumn(Editable = false)]
     public string ComponentConfigAddress;
 }